test(changelog): add make changelog-preview to render latest entry
Export changelog.Parse so a local CHANGELOG.md can be rendered through the
real teaser path instead of a hand-rolled copy. The opt-in preview test
(internal/clawker, gated on CLAWKER_PREVIEW_CHANGELOG) parses the repo
CHANGELOG.md and feeds the newest entry to printChangelogTeaser — the same
renderer a user hits post-upgrade — so a release section (alerts, bullets,
code spans) can be eyeballed before shipping. `make changelog-preview` wraps it.
